requirements.txt
streamlit>=1.30.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pymysql>=1.1.0
python-dotenv>=1.0.0
cryptography>=42.0.0
